Created README.md with Sublime Text instructions
Hi!

I found this repository helpful and got my Sublime Text syntax hilighting working with it. It had many steps, so I decided to document it for the sake of other ST users.

Since I do not have textmate myself and I assume this repo has been created with it in mind, this pull request is only a first step in creating instructions for it too.

Thanks!